+++ CHANGELOG.md
@@ -310,6 +310,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `\OxidEsales\EshopCommunity\Application\Controller\Admin\AdminDetailsController::_generateTextEditor()`, use generateTextEditor()
 - `\OxidEsales\EshopCommunity\Application\Model\Order::getFormattedeliveryCost()`, use getFormattedDeliveryCost()
 - `\OxidEsales\EshopCommunity\Application\Model\Delivery::_isForArticle`, use isDeliveryRuleFitByArticle()
+- `\OxidEsales\EshopCommunity\Application\Model\Voucher::_getGenericDiscoutValue()`, use getGenericDiscountValue()
+- `\OxidEsales\EshopCommunity\Application\Model\Voucher::_getProductDiscoutValue()`, use getProductDiscountValue()
+- `\OxidEsales\EshopCommunity\Application\Model\Voucher::_getCategoryDiscoutValue()`, use getCategoryDiscountValue()
 
 ### Fixed
 - Fix not working actions and promotions [#0005526](https://bugs.oxid-esales.com/view.php?id=5526)